xcpMaxProgCto should be of bigger type, otherwise xcpMaxProgCto > XCPLOADER_PACKET_SIZE_MAX(255) makes no sense

+++ Host/Source/LibOpenBLT/xcploader.c
@@ -127,7 +127,7 @@ static uint16_t xcpMaxCto;
 /** \brief The max number of bytes in the command transmit object (master->slave) during
  *         a programming session.
  */
-static uint8_t xcpMaxProgCto;
+static uint16_t xcpMaxProgCto;
 
 /** \brief The max number of bytes in the data transmit object (slave->master). */
 static uint16_t xcpMaxDto;
